pyproject.toml
[project]
name = "binance-open-interest-downloader"
version = "0.1.0"
description = "Binance Futures Open Interest Data Downloader with scheduling, error handling and SQLite storage"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "requests>=2.31.0",
    "PySocks>=1.7.1",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0"
]

[project.scripts]
binance-oi-downloader = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["binance_open_interest"]
